transaksi, upload, validasi, terima potongan cms

+++ app/Http/Controllers/UserController.php
@@ -3,9 +3,13 @@
 namespace App\Http\Controllers;
 
 use App\Http\Requests\UserRequest;
+use Exception;
+use Illuminate\Auth\Events\Validated;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Route;
 
 class UserController extends Controller
 {
@@ -25,6 +29,9 @@ class UserController extends Controller
             'daftar_skpd' => DB::table('ms_skpd')->get()
         ];
 
+        $hak = cek_akses();
+
+        if ($hak == '0') abort(403);
         return view('master.user.create')->with($data);
     }
 
@@ -46,11 +53,15 @@ class UserController extends Controller
 
     public function show($id)
     {
-        $user = DB::table('user')->where('id', $id)->first();
+        $user = DB::table('user')->where('id', decrypt($id))->first();
         $data = [
             'user' => $user,
             'role' => DB::table('role')->where('id', $user->role)->first()
         ];
+
+        $hak = cek_akses();
+
+        if ($hak == '0') abort(403);
 
         return view('master.user.show')->with($data);
     }
@@ -64,6 +75,10 @@ class UserController extends Controller
             'daftar_skpd' => DB::table('ms_skpd')->get(),
             'role' => DB::table('role')->where('id', $user->role)->first(),
         ];
+
+        $hak = cek_akses();
+
+        if ($hak == '0') abort(403);
 
         return view('master.user.edit')->with($data);
     }
@@ -85,6 +100,9 @@ class UserController extends Controller
 
     public function destroy($id)
     {
+        $hak = cek_akses();
+
+        if ($hak == '0') abort(403);
         DB::table('user')->where('id', $id)->delete();
         return redirect()->route('user.index');
     }
